Started to clean up misc. code used for testing, moved various functions out of VRAY_cluserThis.C into other source files, etc.
