+++ Doxyfile
@@ -31,7 +31,7 @@ PROJECT_NAME           = clusterThis
 # This could be handy for archiving the generated documentation or
 # if some version control system is used.
 
-PROJECT_NUMBER         = 2.1.337
+PROJECT_NUMBER         = 2.5.338
 
 # The OUTPUT_DIRECTORY tag is used to specify the (relative or absolute)
 # base path where the generated documentation will be put.
@@ -583,6 +583,7 @@ WARN_LOGFILE           =
 INPUT                  = VRAY_clusterThis.h \
                          VRAY_clusterThisInstance.C \
                          VRAY_clusterUtil.C \
+                         VRAY_clusterThisParms.C \
                          VRAY_clusterCVEXUtil.C \
                          VRAYprocedural \
                          Doxyfile \
