+++ VRAY_clusterThis.C
@@ -1,6 +1,6 @@
 /* ******************************************************************************
 *
-* clusterThis mantra DSO for render time geo clustering
+* clusterThis mantra DSO for render time geo instancing/clustering
 *
 * Description : This DSO  will instance geomtery or volumes during the render of mantra IFD's
 *
@@ -12,36 +12,7 @@
 *
 ***************************************************************************** */
 
-///////////////////////////////////////////////////////////////////////////
-//
-// Copyright (c) 2012 DreamWorks Animation LLC
-//
-// All rights reserved. This software is distributed under the
-// Mozilla Public License 2.0 ( http://www.mozilla.org/MPL/2.0/ )
-//
-// Redistributions of source code must retain the above copyright
-// and license notice and the following restrictions and disclaimer.
-//
-// *     Neither the name of DreamWorks Animation nor the names of
-// its contributors may be used to endorse or promote products derived
-// from this software without specific prior written permission.
-//
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
-// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
-// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
-// A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
-// OWNER OR CONTRIBUTORS BE LIABLE FOR ANY INDIRECT, INCIDENTAL,
-// SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
-// LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
-// DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
-// THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-// IN NO EVENT SHALL THE COPYRIGHT HOLDERS' AND CONTRIBUTORS' AGGREGATE
-// LIABILITY FOR ALL CLAIMS REGARDLESS OF THEIR BASIS EXCEED US$250.00.
-//
-///////////////////////////////////////////////////////////////////////////
-//
+// Open VDB copyright information can be found here: openvdb.org
 
 #ifndef __VRAY_clusterThis_C__
 #define __VRAY_clusterThis_C__
@@ -107,6 +78,7 @@ namespace hutil = houdini_utils;
 
 #include "version.h"
 #include "VRAY_clusterThis.h"
+#include "VRAY_clusterThisParms.C"
 #include "VRAY_clusterThisUtil.C"
 #include "VRAY_clusterThisRender.C"
 #include "VRAY_clusterThisInstance.C"
@@ -117,121 +89,6 @@ namespace hutil = houdini_utils;
 
 class VRAY_clusterThis_Exception;
 
-
-
-/* ******************************************************************************
-*  Function Name : theArgs()
-*
-*  Description : Data for the arguments...
-*
-*  Input Arguments : None
-*
-*  Return Value :
-*
-***************************************************************************** */
-static VRAY_ProceduralArg theArgs[] = {
-
-   VRAY_ProceduralArg("prim_type", "integer", "0"),
-   VRAY_ProceduralArg("use_geo_file", "integer", "0"),
-   VRAY_ProceduralArg("src_geo_file", "string", "default.bgeo"),
-   VRAY_ProceduralArg("num_copy", "integer", "10"),
-   VRAY_ProceduralArg("recursion", "integer", "2"),
-   VRAY_ProceduralArg("radius",    "real",  "0.1"),
-   VRAY_ProceduralArg("size",  "real", "0.01 0.01 0.01"),
-   VRAY_ProceduralArg("freq",    "real",  "1.0 1.0 1.0"),
-   VRAY_ProceduralArg("offset",    "real",  "0.0 0.0 0.0"),
-   VRAY_ProceduralArg("birth_prob", "real",  "0.5"),
-   VRAY_ProceduralArg("motion_blur", "integer", "0"),
-   VRAY_ProceduralArg("backtrack_mb", "integer", "0"),
-   VRAY_ProceduralArg("mb_shutter", "real", "0.1"),
-   VRAY_ProceduralArg("mb_shutter2", "real", "0.9"),
-   VRAY_ProceduralArg("verbose", "integer", "0"),
-   VRAY_ProceduralArg("time", "real",  "0.0"),
-   VRAY_ProceduralArg("noise_type", "integer", "0"),
-   VRAY_ProceduralArg("noise_amp", "real", "0.1"),
-   VRAY_ProceduralArg("noise_rough", "real", "0.1"),
-   VRAY_ProceduralArg("noise_atten", "real", "0.0"),
-   VRAY_ProceduralArg("noise_seed", "integer", "7"),
-   VRAY_ProceduralArg("noise_fractal_depth", "integer", "3"),
-   VRAY_ProceduralArg("geo_file", "string", "default.bgeo"),
-   VRAY_ProceduralArg("filter_type", "integer", "0"),
-   VRAY_ProceduralArg("filter_amp", "real", "0.0"),
-   VRAY_ProceduralArg("temp_file_path", "string", "/tmp/geo/"),
-   VRAY_ProceduralArg("temp_file", "integer", "0"),
-   VRAY_ProceduralArg("save_temp_file", "integer", "0"),
-   VRAY_ProceduralArg("otl_version", "string", "DCA_VERSION"),
-
-   VRAY_ProceduralArg("CVEX_shader", "string", "default.vex"),
-   VRAY_ProceduralArg("CVEX_exec", "integer", "0"),
-   VRAY_ProceduralArg("CVEX_shader_prim", "string", "default.vex"),
-   VRAY_ProceduralArg("CVEX_exec_prim", "integer", "0"),
-   VRAY_ProceduralArg("CVEX_shader_pre", "string", "default.vex"),
-   VRAY_ProceduralArg("CVEX_exec_pre", "integer", "0"),
-   VRAY_ProceduralArg("CVEX_shader_post", "string", "default.vex"),
-   VRAY_ProceduralArg("CVEX_exec_post", "integer", "0"),
-
-   VRAY_ProceduralArg("Cd_pt", "integer", "0"),
-   VRAY_ProceduralArg("Alpha_pt", "integer", "0"),
-   VRAY_ProceduralArg("v_pt", "integer", "0"),
-   VRAY_ProceduralArg("N_pt", "integer", "0"),
-   VRAY_ProceduralArg("pscale_pt", "integer", "0"),
-
-   VRAY_ProceduralArg("Cd_prim", "integer", "0"),
-   VRAY_ProceduralArg("Alpha_prim", "integer", "0"),
-   VRAY_ProceduralArg("v_prim", "integer", "0"),
-   VRAY_ProceduralArg("N_prim", "integer", "0"),
-   VRAY_ProceduralArg("pscale_prim", "integer", "0"),
-   VRAY_ProceduralArg("weight_prim", "integer", "0"),
-   VRAY_ProceduralArg("width_prim", "integer", "0"),
-
-   VRAY_ProceduralArg("cvex_Cd_pt", "integer", "0"),
-   VRAY_ProceduralArg("cvex_Alpha_pt", "integer", "0"),
-   VRAY_ProceduralArg("cvex_v_pt", "integer", "0"),
-   VRAY_ProceduralArg("cvex_N_pt", "integer", "0"),
-   VRAY_ProceduralArg("cvex_pscale_pt", "integer", "0"),
-
-   VRAY_ProceduralArg("cvex_Cd_prim", "integer", "0"),
-   VRAY_ProceduralArg("cvex_Alpha_prim", "integer", "0"),
-   VRAY_ProceduralArg("cvex_v_prim", "integer", "0"),
-   VRAY_ProceduralArg("cvex_N_prim", "integer", "0"),
-   VRAY_ProceduralArg("cvex_pscale_prim", "integer", "0"),
-   VRAY_ProceduralArg("cvex_weight_prim", "integer", "0"),
-   VRAY_ProceduralArg("cvex_width_prim", "integer", "0"),
-
-   VRAY_ProceduralArg("post_process", "integer", "1"),
-   VRAY_ProceduralArg("nn_post_process", "integer", "1"),
-   VRAY_ProceduralArg("nn_post_pos_influence", "real", "0.1"),
-   VRAY_ProceduralArg("nn_post_vel_influence", "real", "0.1"),
-   VRAY_ProceduralArg("vdb_post_process", "integer", "1"),
-   VRAY_ProceduralArg("vdb_post_raster_type", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_ws_units", "integer", "1"),
-   VRAY_ProceduralArg("vdb_post_fog_volume", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_dx", "real", "1.0"),
-   VRAY_ProceduralArg("vdb_post_gradient_width", "real", "0.5"),
-   VRAY_ProceduralArg("vdb_post_voxel_size", "real", "0.025"),
-   VRAY_ProceduralArg("vdb_post_radius_min", "real", "1.5"),
-   VRAY_ProceduralArg("vdb_post_radius_mult", "real", "1.0"),
-   VRAY_ProceduralArg("vdb_post_velocity_mult", "real", "1.0"),
-   VRAY_ProceduralArg("vdb_post_bandwidth", "real", "0.2"),
-   VRAY_ProceduralArg("vdb_post_falloff", "real", "0.5"),
-   VRAY_ProceduralArg("vdb_post_pos_influence", "real", "0.1"),
-   VRAY_ProceduralArg("vdb_post_vel_influence", "real", "0.1"),
-   VRAY_ProceduralArg("vdb_post_normal_influence", "real", "0.1"),
-   VRAY_ProceduralArg("vdb_post_median_filter", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_mean_filter", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_mean_curvature_filter", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_laplacian_filter", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_median_iterations", "integer", "4"),
-   VRAY_ProceduralArg("vdb_post_mean_iterations", "integer", "4"),
-   VRAY_ProceduralArg("vdb_post_mean_curvature_iterations", "integer", "4"),
-   VRAY_ProceduralArg("vdb_post_laplacian_iterations", "integer", "4"),
-   VRAY_ProceduralArg("vdb_post_offset_filter", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_offset_filter_amount", "real", "0.1"),
-   VRAY_ProceduralArg("vdb_post_renormalize_filter", "integer", "0"),
-   VRAY_ProceduralArg("vdb_post_write_debug_file", "integer", "0"),
-
-   VRAY_ProceduralArg()
-};
 
 
 /* ******************************************************************************
@@ -294,231 +151,6 @@ VRAY_clusterThis_Exception::VRAY_clusterThis_Exception(std::string msg, int code
 
 //   };
 
-
-
-
-/* ******************************************************************************
-*  Function Name : exitClusterThis()
-*
-*  Description :  Start the exit process
-*
-*  Input Arguments : void *data
-*
-*  Return Value :
-*
-***************************************************************************** */
-void VRAY_clusterThis::exitClusterThis(void * data)
-{
-   VRAY_clusterThis * me = (VRAY_clusterThis *)data;
-
-   if(me->myVerbose > CLUSTER_MSG_INFO)
-      std::cout << std::endl << std::endl << "VRAY_clusterThis::exitClusterThis() - Preparing to exit!" << std::endl;
-
-
-//   if(me->myVerbose > CLUSTER_MSG_INFO)
-//      cout << "VRAY_clusterThis::exitClusterThis() myTempFname: " << (const char *)me->myTempFname << endl;
-
-
-//   if(me->tempFileDeleted) {
-//         me->tempFileDeleted = true;
-//         cout << "VRAY_clusterThis::exitClusterThis(): " << me->tempFileDeleted << endl;
-////         me->exitClusterThisReal((const char *)me->myTempFname);
-//         me->exitClusterThisReal(data);
-//      }
-
-
-   if(me->myVerbose > CLUSTER_MSG_INFO)
-      cout << "VRAY_clusterThis::exitClusterThis() - Running exit processing" << endl;
-
-//   cout << "VRAY_clusterThis::exitClusterThis(): " << me->tempFileDeleted << endl;
-//
-//   cout << "VRAY_clusterThis::exitClusterThis() - temp filename " << me->myTempFname << endl;
-
-//   const char * fname = me->myTempFname;
-//
-////   ofstream myStream;
-//
-////   myStream.open("exit_data.txt", ios_base::app);
-////   myStream << this->exitData.exitTime << std::endl;
-////   myStream << this->exitData.exitCode << std::endl;
-////   myStream.flush();
-////   myStream.close();
-////   cout << "VRAY_clusterThis::exitClusterThisReal() : " << this->exitData.exitTime << endl;
-////
-//
-//   struct stat fileResults;
-//
-//   if(me->myUseTempFile && !me->mySaveTempFile) {
-//         if((UT_String(fname)).isstring() && stat(fname, &fileResults) == 0) {
-//               if(me->myVerbose > CLUSTER_MSG_INFO)
-//                  cout << "VRAY_clusterThis::exitClusterThis() - Found temp file " << fname << endl;
-//               if(!remove(fname) && (me->myVerbose > CLUSTER_MSG_INFO))
-//                  cout << "VRAY_clusterThis::exitClusterThis() - Removed geometry temp file: " << fname << endl;
-//            }
-//         else
-//            if(me->myVerbose > CLUSTER_MSG_INFO)
-//               cout << "VRAY_clusterThis::exitClusterThis() - Did not find temp file " << fname << endl << endl;
-//      }
-//
-//
-
-   const char * fname = me->myTempFname;
-
-   me->exitClusterThisReal(fname);
-//         me->exitClusterThisReal(data);
-
-}
-
-
-
-/* ******************************************************************************
-*  Function Name : exitClusterThisReal()
-*
-*  Description : Clean up temp file if used, save stats to file and DB
-*
-*  Input Arguments : None
-*
-*  Return Value :
-*
-* ***************************************************************************** */
-//void VRAY_clusterThis::exitClusterThisReal(void * data)
-void VRAY_clusterThis::exitClusterThisReal(const char * fname)
-{
-   struct stat fileResults;
-//   VRAY_clusterThis * me = (VRAY_clusterThis *)data;
-
-
-   if(myVerbose > CLUSTER_MSG_INFO)
-      cout << "VRAY_clusterThis::exitClusterThisReal() - Running exit processing" << endl;
-
-
-//   cout << "VRAY_clusterThis::exitClusterThisReal(): " << tempFileDeleted << endl;
-//
-//   cout << "VRAY_clusterThis::exitClusterThisReal() - temp filename " << myTempFname << endl;
-//
-//   const char * fname = me->myTempFname;
-
-//   ofstream myStream;
-
-//   myStream.open("exit_data.txt", ios_base::app);
-//   myStream << this->exitData.exitTime << std::endl;
-//   myStream << this->exitData.exitCode << std::endl;
-//   myStream.flush();
-//   myStream.close();
-//   cout << "VRAY_clusterThis::exitClusterThisReal() : " << this->exitData.exitTime << endl;
-//
-
-
-   if(this->myUseTempFile && !this->mySaveTempFile) {
-         if((UT_String(fname)).isstring() && stat(fname, &fileResults) == 0) {
-               if(myVerbose > CLUSTER_MSG_INFO)
-                  cout << "VRAY_clusterThis::exitClusterThisReal() - Found temp file " << fname << endl;
-               if(!remove(fname) && (myVerbose > CLUSTER_MSG_INFO))
-                  cout << "VRAY_clusterThis::exitClusterThisReal() - Removed geometry temp file: " << fname << endl;
-            }
-         else
-            if(myVerbose > CLUSTER_MSG_INFO)
-               cout << "VRAY_clusterThis::exitClusterThisReal() - Did not find temp file " << fname << endl << endl;
-      }
-
-
-
-   if(myVerbose > CLUSTER_MSG_INFO)
-      std::cout << "VRAY_clusterThis::exitClusterThisReal() - Exiting" << std::endl;
-}
-
-
-
-
-
-/* ******************************************************************************
-*  Function Name : convert()
-*
-*  Description : convert point cloud to VDB level set or fog volume
-*
-*  Input Arguments : None
-*
-*  Return Value :
-*
-* ***************************************************************************** */
-void VRAY_clusterThis::convert(
-   openvdb::ScalarGrid::Ptr outputGrid,
-   ParticleList & paList,
-   const Settings & settings,
-   hvdb::Interrupter & boss)
-{
-
-   openvdb::tools::ParticlesToLevelSet<openvdb::ScalarGrid, ParticleList, hvdb::Interrupter> raster(*outputGrid, boss);
-
-//   std::cout << "VRAY_clusterThis::convert() " << std::endl;
-
-
-   raster.setRmin(settings.mRadiusMin);
-
-   if(myVerbose == CLUSTER_MSG_DEBUG) {
-         std::cout << "VRAY_clusterThis::convert(): raster.getVoxelSize(): " << raster.getVoxelSize() << std::endl;
-         std::cout << "VRAY_clusterThis::convert(): raster.getRmin(): " << raster.getRmin() << std::endl;
-         std::cout << "VRAY_clusterThis::convert(): raster.getHalfWidth(): " << raster.getHalfWidth() << std::endl;
-      }
-
-   if(raster.getHalfWidth() < openvdb::Real(2)) {
-         std::cout << "VRAY_clusterThis::convert(): Half width of narrow-band < 2 voxels which creates holes when meshed!" << std::endl;
-      }
-   else
-      if(raster.getHalfWidth() > openvdb::Real(1000)) {
-            throw std::runtime_error(
-               "VRAY_clusterThis::convert(): Half width of narrow-band > 1000 voxels which exceeds memory limitations!");
-         }
-
-   if(settings.mRasterizeTrails && paList.hasVelocity()) {
-         if(myVerbose == CLUSTER_MSG_DEBUG)
-            std::cout << "VRAY_clusterThis::convert(): rasterizing trails"  << std::endl;
-         raster.rasterizeTrails(paList, settings.mDx);
-      }
-   else {
-         if(myVerbose == CLUSTER_MSG_DEBUG)
-            std::cout << "VRAY_clusterThis::convert(): rasterizing spheres"  << std::endl;
-         raster.rasterizeSpheres(paList);
-      }
-
-   if(boss.wasInterrupted()) {
-         if(myVerbose == CLUSTER_MSG_DEBUG)
-            std::cout << "VRAY_clusterThis::convert(): Process was interrupted"  << std::endl;
-         return;
-      }
-
-   // Convert the level-set into a fog volume.
-   if(settings.mFogVolume) {
-         if(myVerbose == CLUSTER_MSG_DEBUG)
-            std::cout << "VRAY_clusterThis::convert(): converting to fog volume"  << std::endl;
-         float cutOffDist = std::numeric_limits<float>::max();
-         if(settings.mGradientWidth > 1e-6)
-            cutOffDist = settings.mGradientWidth;
-         openvdb::tools::levelSetToFogVolume(*outputGrid, cutOffDist, false);
-      }
-
-// print stats of the vdb grid
-   if(myVerbose == CLUSTER_MSG_DEBUG)
-      outputGrid->print();
-
-}
-
-
-
-
-int VRAY_clusterThis::convertVDBUnits()
-{
-//     const bool toWSUnits = static_cast<bool>(evalInt("worldSpaceUnits", 0, 0));
-//
-//     if (toWSUnits) {
-//         setFloat("bandWidthWS", 0, 0, evalFloat("bandWidth", 0, 0) * mVoxelSize);
-//         return 1;
-//     }
-//
-//     setFloat("bandWidth", 0, 0, evalFloat("bandWidthWS", 0, 0) / mVoxelSize);
-//
-   return 1;
-}
 
 
 
@@ -952,380 +584,6 @@ int VRAY_clusterThis::initialize(const UT_BoundingBox * box)
 
 
 /* ******************************************************************************
-*  Function Name : getOTLParameters
-*
-*  Description : Get OTL parameters
-*
-*  Input Arguments : None
-*
-*  Return Value : None
-*
-***************************************************************************** */
-int VRAY_clusterThis::getOTLParameters()
-{
-
-//   UT_String geo_fname, temp_fname, cvex_fname, otl_version;
-   const int * int_ptr;
-   const fpreal * flt_ptr;
-   const char ** char_handle;
-
-   if(int_ptr = VRAY_Procedural::getIParm("prim_type"))
-      myPrimType = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("num_copy"))
-      myNumCopies = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("recursion"))
-      myRecursion = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("motion_blur"))
-      myDoMotionBlur = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("backtrack_mb"))
-      myUseBacktrackMB = *int_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("mb_shutter"))
-      myShutter = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("mb_shutter2"))
-      myShutter2 = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("radius"))
-      myRadius = *flt_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("noise_type"))
-      myNoiseType = *int_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("freq")) {
-         myFreqX = *flt_ptr++;
-         myFreqY = *flt_ptr++;
-         myFreqZ = *flt_ptr;
-      }
-
-   if(flt_ptr = VRAY_Procedural::getFParm("offset")) {
-         myOffsetX = *flt_ptr++;
-         myOffsetY = *flt_ptr++;
-         myOffsetZ = *flt_ptr;
-      }
-
-   if(flt_ptr = VRAY_Procedural::getFParm("noise_rough"))
-      myRough = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("size")) {
-         mySize[0] = *flt_ptr++;
-         mySize[1] = *flt_ptr++;
-         mySize[2] = *flt_ptr;
-      }
-
-   if(flt_ptr = VRAY_Procedural::getFParm("birth_prob"))
-      myBirthProb = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("time"))
-      myCurrentTime = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("noise_amp"))
-      myNoiseAmp = *flt_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("filter_type"))
-      myFilterType = *int_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("filter_amp"))
-      myFilterAmp = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("noise_atten"))
-      myNoiseAtten = *flt_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("noise_seed"))
-      myNoiseSeed = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("noise_fractal_depth"))
-      myFractalDepth = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("use_geo_file"))
-      myUseGeoFile = *int_ptr;
-
-   if(char_handle = VRAY_Procedural::getSParm("src_geo_file")) {
-         mySrcGeoFname = (UT_String)(*char_handle);
-         mySrcGeoFname.harden();
-      }
-
-
-   if(char_handle = VRAY_Procedural::getSParm("geo_file")) {
-         myGeoFile = (UT_String)(*char_handle);
-         myGeoFile.harden();
-      }
-
-   if(char_handle = VRAY_Procedural::getSParm("temp_file_path")) {
-         myTempFname = (UT_String)(*char_handle);
-         myTempFname.harden();
-      }
-
-
-   if(int_ptr = VRAY_Procedural::getIParm("temp_file"))
-      myUseTempFile = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("save_temp_file"))
-      mySaveTempFile =  *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("verbose"))
-      myVerbose = *int_ptr;
-
-//   if(char_handle = VRAY_Procedural::getSParm("otl_version")) {
-//         myOTLVersion = (UT_String)(*char_handle);
-//         myOTLVersion.harden();
-//      }
-
-
-
-
-   if(char_handle = VRAY_Procedural::getSParm("CVEX_shader")) {
-         myCVEXFname = (UT_String)(*char_handle);
-         myCVEXFname.harden();
-      }
-
-
-   if(int_ptr = VRAY_Procedural::getIParm("CVEX_exec"))
-      myCVEX_Exec =  *int_ptr;
-
-
-   if(char_handle = VRAY_Procedural::getSParm("CVEX_shader_prim")) {
-         myCVEXFname_prim = (UT_String)(*char_handle);
-         myCVEXFname_prim.harden();
-      }
-
-   if(int_ptr = VRAY_Procedural::getIParm("CVEX_exec_prim"))
-      myCVEX_Exec_prim = *int_ptr;
-
-
-   if(char_handle = VRAY_Procedural::getSParm("CVEX_shader_pre")) {
-         myCVEXFname_pre = (UT_String)(*char_handle);
-         myCVEXFname_pre.harden();
-      }
-
-   if(int_ptr = VRAY_Procedural::getIParm("CVEX_exec_pre"))
-      myCVEX_Exec_pre = *int_ptr;
-
-   if(char_handle = VRAY_Procedural::getSParm("CVEX_shader_post")) {
-         myCVEXFname_post = (UT_String)(*char_handle);
-         myCVEXFname_post.harden();
-      }
-
-   if(int_ptr = VRAY_Procedural::getIParm("CVEX_exec_post"))
-      myCVEX_Exec_post = *int_ptr;
-
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_Cd_pt"))
-      myCVEXPointVars.cvex_Cd_pt = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_Alpha_pt"))
-      myCVEXPointVars.cvex_Alpha_pt = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_v_pt"))
-      myCVEXPointVars.cvex_v_pt = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_N_pt"))
-      myCVEXPointVars.cvex_N_pt = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_pscale_pt"))
-      myCVEXPointVars.cvex_pscale_pt = *int_ptr;
-
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_Cd_prim"))
-      myCVEXPrimVars.cvex_Cd_prim = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_Alpha_prim"))
-      myCVEXPrimVars.cvex_Alpha_prim = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_v_prim"))
-      myCVEXPrimVars.cvex_v_prim = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_N_prim"))
-      myCVEXPrimVars.cvex_N_prim = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_pscale_prim"))
-      myCVEXPrimVars.cvex_pscale_prim = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_weight_prim"))
-      myCVEXPrimVars.cvex_weight_prim = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("cvex_width_prim"))
-      myCVEXPrimVars.cvex_width_prim = *int_ptr;
-
-
-   // Post processing parms
-   if(int_ptr = VRAY_Procedural::getIParm("post_process"))
-      myPostProcess = *int_ptr;
-
-   // Nearest neighbor post processing parms
-   if(int_ptr = VRAY_Procedural::getIParm("nn_post_process"))
-      myNNPostProcess = *int_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("nn_post_pos_influence"))
-      myNNPostPosInfluence = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("nn_post_vel_influence"))
-      myNNPostVelInfluence = *flt_ptr;
-
-   // VDB pre processing parms
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_process"))
-      myVDBPostProcess = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_raster_type"))
-      myPostRasterType = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_ws_units"))
-      myPostWSUnits = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_fog_volume"))
-      myPostFogVolume = *int_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_dx"))
-      myPostDx = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_gradient_width"))
-      myPostGradientWidth = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_voxel_size"))
-      myPostVoxelSize = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_radius_min"))
-      myPostRadiusMin = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_radius_mult"))
-      myPostVDBRadiusMult = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_velocity_mult"))
-      myPostVDBVelocityMult = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_bandwidth"))
-      myPostBandWidth = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_falloff"))
-      myPostFalloff = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_pos_influence"))
-      myPostPosInfluence = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_vel_influence"))
-      myPostVelInfluence = *flt_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_normal_influence"))
-      myPostNormalInfluence = *flt_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_median_filter"))
-      myPostVDBMedianFilter = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_mean_filter"))
-      myPostVDBMeanFilter = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_mean_curvature_filter"))
-      myPostVDBMeanCurvatureFilter = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_laplacian_filter"))
-      myPostVDBLaplacianFilter = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_median_iterations"))
-      myPostVDBMedianIterations = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_mean_iterations"))
-      myPostVDBMeanIterations = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_mean_curvature_iterations"))
-      myPostVDBMeanCurvatureIterations = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_laplacian_iterations"))
-      myPostVDBLaplacianIterations = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_offset_filter"))
-      myPostVDBOffsetFilter = *int_ptr;
-
-   if(flt_ptr = VRAY_Procedural::getFParm("vdb_post_offset_filter_amount"))
-      myPostVDBOffsetFilterAmount = *flt_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_renormalize_filter"))
-      myPostVDBReNormalizeFilter = *int_ptr;
-
-   if(int_ptr = VRAY_Procedural::getIParm("vdb_post_write_debug_file"))
-      myPostVDBWriteDebugFiles = *int_ptr;
-
-
-   return 0;
-
-}
-
-/* ******************************************************************************
-*  Function Name : getBoundingBox
-*
-*  Description :  Get the bounding box for this VRAY_clusterThis object
-*
-*  Input Arguments : None
-*
-*  Return Value : None
-*
-***************************************************************************** */
-void VRAY_clusterThis::getBoundingBox(UT_BoundingBox & box)
-{
-
-   box = myVelBox;
-
-#ifdef DEBUG
-   std::cout << "VRAY_clusterThis::getBoundingBox() box: " << box << std::endl;
-#endif
-
-}
-
-
-/* ******************************************************************************
-*  Function Name : checkRequiredAttributes
-*
-*  Description :  Check that all the required attributes are in the point cloud
-*
-*  Input Arguments : None
-*
-*  Return Value : None
-*
-***************************************************************************** */
-void VRAY_clusterThis::checkRequiredAttributes()
-{
-
-   // TODO: Confirm that all required attrs are being checked for each instance type
-
-//   std::cout << "VRAY_clusterThis::checkRequiredAttributes()" << std::endl;
-   // Check for required attributes
-   if(myPointAttrRefs.Cd.isInvalid()) {
-         cout << "Incoming points must have Cd attribute! Throwing exception ..." << endl;
-         throw VRAY_clusterThis_Exception("VRAY_clusterThis::render() Incoming points must have Cd attribute! ", 1);
-      }
-
-   if(myPointAttrRefs.Alpha.isInvalid()) {
-         cout << "Incoming points must have Alpha attribute! Throwing exception ..." << endl;
-         throw VRAY_clusterThis_Exception("VRAY_clusterThis::render() Incoming points must have Alpha attribute! ", 1);
-      }
-
-   if(myPointAttrRefs.v.isInvalid()) {
-         cout << "Incoming points must have v attribute! Throwing exception ..." << endl;
-         throw VRAY_clusterThis_Exception("VRAY_clusterThis::render() Incoming points must have v attribute! ", 1);
-      }
-
-   if(myPointAttrRefs.N.isInvalid()) {
-         cout << "Incoming points must have N attribute! Throwing exception ..." << endl;
-         throw VRAY_clusterThis_Exception("VRAY_clusterThis::render() Incoming points must have N attribute! ", 1);
-      }
-
-   if(myPointAttrRefs.pscale.isInvalid()) {
-         cout << "Incoming points must have pscale attribute! Throwing exception ..." << endl;
-         throw VRAY_clusterThis_Exception("VRAY_clusterThis::render() Incoming points must have pscale attribute! ", 1);
-      }
-
-   if(myPointAttrRefs.id.isInvalid()) {
-         cout << "Incoming points must have id attribute Throwing exception ..." << endl;
-         throw VRAY_clusterThis_Exception("VRAY_clusterThis::render() Incoming points must have id attribute! ", 1);
-      }
-
-}
-
-
-/* ******************************************************************************
 *  Function Name : calculateNewPosition()
 *
 *  Description :   Calculate the position of the new instance
@@ -1415,133 +673,10 @@ inline void VRAY_clusterThis::calculateNewPosition(fpreal theta, uint32 i, uint3
 
 
 /* ******************************************************************************
-*  Function Name : dumpParameters()
-*
-*  Description :   Dump the user parameters to the console
-*
-*  Input Arguments : None
-*
-*  Return Value : None
-*
-***************************************************************************** */
-void VRAY_clusterThis::dumpParameters()
-{
-
-   std::cout << std::endl;
-
-   std::cout << "VRAY_clusterThis::dumpParameters() **** SETUP PARAMETERS ****" << std::endl;
-
-   std::cout << "VRAY_clusterThis::dumpParameters() myPrimType: " << myPrimType << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myNumCopies: " << myNumCopies << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myUseGeoFile: " << myUseGeoFile << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() mySrcGeoFname: " << mySrcGeoFname << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myRecursion: " << myRecursion << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myRadius: " << myRadius << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myRough: " << myRough << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() mySize: " << mySize[0] << " " << mySize[1] << " " << mySize[2] << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myFreq: " << myFreqX << " " << myFreqY << " " << myFreqZ << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myOffset: " << myOffsetX << " " << myOffsetY << " " << myOffsetZ << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myBirthProb: " << myBirthProb << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCurrentTime: " << myCurrentTime << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myDoMotionBlur: " << myDoMotionBlur << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myShutter: " << myShutter << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myVerbose: " << myVerbose << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myGeoFile: " << myGeoFile << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myTempFname: " << myTempFname << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myUseTempFile: " << myUseTempFile << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() mySaveTempFile: " << mySaveTempFile << std::endl;
-
-
-   std::cout << "VRAY_clusterThis::dumpParameters() **** NOISE PARAMETERS ****" << std::endl;
-
-   std::cout << "VRAY_clusterThis::dumpParameters() myNoiseType: " << myNoiseType << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myNoiseAmp: " << myNoiseAmp << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myNoiseAtten: " << myNoiseAtten << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myNoiseSeed: " << myNoiseSeed << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myFractalDepth: " << myFractalDepth << std::endl;
-
-
-   std::cout << "VRAY_clusterThis::dumpParameters() **** CVEX PARAMETERS ****" << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXFname: " << myCVEXFname << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEX_Exec: " << myCVEX_Exec << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXFname_prim: " << myCVEXFname_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEX_Exec_prim: " << myCVEX_Exec_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXFname_pre: " << myCVEXFname_pre << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEX_Exec_pre: " << myCVEX_Exec_pre << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXFname_post: " << myCVEXFname_post << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEX_Exec_post: " << myCVEX_Exec_post << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPointVars.cvex_Cd_pt: " << myCVEXPointVars.cvex_Cd_pt << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPointVars.cvex_Alpha_pt: " << myCVEXPointVars.cvex_Alpha_pt << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPointVars.cvex_N_pt: " << myCVEXPointVars.cvex_Alpha_pt << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPointVars.cvex_v_pt: " << myCVEXPointVars.cvex_v_pt << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPointVars.cvex_pscale_pt: " << myCVEXPointVars.cvex_pscale_pt << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPrimVars.cvex_Cd_prim: " << myCVEXPrimVars.cvex_Cd_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPrimVars.cvex_Alpha_prim: " << myCVEXPrimVars.cvex_Alpha_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPrimVars.cvex_N_prim: " << myCVEXPrimVars.cvex_N_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPrimVars.cvex_v_prim: " << myCVEXPrimVars.cvex_v_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPrimVars.cvex_pscale_prim: " << myCVEXPrimVars.cvex_pscale_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPrimVars.cvex_weight_prim: " << myCVEXPrimVars.cvex_weight_prim << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myCVEXPrimVars.cvex_width_prim: " << myCVEXPrimVars.cvex_width_prim << std::endl;
-
-   std::cout << "VRAY_clusterThis::dumpParameters() **** POST PROCESS PARAMETERS ****" << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostProcess: " << myPostProcess << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myNNPostProcess: " << myNNPostProcess << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myNNPostPosInfluence: " << myNNPostPosInfluence << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myNNPostVelInfluence: " << myNNPostVelInfluence << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myVDBPostProcess: " << myVDBPostProcess << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostRasterType: " << myPostRasterType << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostDx: " << myPostDx << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostFogVolume: " << myPostFogVolume << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostRadiusMin: " << myPostRadiusMin << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBVelocityMult: " << myPostVDBVelocityMult << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBRadiusMult: " << myPostVDBRadiusMult << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostGradientWidth: " << myPostGradientWidth << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVoxelSize: " << myPostVoxelSize << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostBandWidth: " << myPostBandWidth << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostWSUnits: " << myPostWSUnits << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostFalloff: " << myPostFalloff << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostPosInfluence: " << myPostPosInfluence << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVelInfluence: " << myPostVelInfluence << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostNormalInfluence: " << myPostNormalInfluence << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBMedianFilter: " << myPostVDBMedianFilter << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBMeanFilter: " << myPostVDBMeanFilter << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBMeanCurvatureFilter: " << myPostVDBMeanCurvatureFilter << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBLaplacianFilter: " << myPostVDBLaplacianFilter << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBMedianIterations: " << myPostVDBMedianIterations << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBMeanIterations: " << myPostVDBMeanIterations << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBMeanCurvatureIterations: " << myPostVDBMeanCurvatureIterations << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBLaplacianIterations: " << myPostVDBLaplacianIterations << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBOffsetFilter: " << myPostVDBOffsetFilter << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBOffsetFilterAmount: " << myPostVDBOffsetFilterAmount << std::endl;
-
-//   std::cout << "VRAY_clusterThis::dumpParameters() myVDBReNormalizeFilter: " << myVDBReNormalizeFilter << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myPostVDBWriteDebugFiles: " << myPostVDBWriteDebugFiles << std::endl;
-
-   std::cout << "VRAY_clusterThis::dumpParameters() **** MISC PARAMETERS ****" << std::endl;
-//   std::cout << "VRAY_clusterThis::dumpParameters() myOTLVersion: " << myOTLVersion << std::endl;
-   std::cout << "VRAY_clusterThis::dumpParameters() myVelocityScale: " << myVelocityScale << std::endl;
-//   std::cout << "VRAY_clusterThis::dumpParameters() myFilterType: " << myFilterType << std::endl;
-//   std::cout << "VRAY_clusterThis::dumpParameters() myFilterAmp: " << myFilterAmp << std::endl;
-//   std::cout << "VRAY_clusterThis::dumpParameters() myCopyAttrs: " << myCopyAttrs << std::endl;
-//   std::cout << "VRAY_clusterThis::dumpParameters() myBlendAttrs: " << myBlendAttrs << std::endl;
-
-   std::cout << std::endl;
-
-
-}
-
-
-/* ******************************************************************************
 *  Function Name : preLoadGeoFile()
 *
 *  Description :
 *
-
-
-
-
-
-
 *  Input Arguments : GU_Detail *file_gdp
 *
 *  Return Value : file load status
