+++ VRAY_clusterThisRender.C
@@ -2,23 +2,8 @@
 *
 *  VRAY_clusterThisRender
 *
-* $RCSfile: VRAY_clusterThisRender.C,v $
 *
 * Description :
-*
-* $Revision: 1.17 $
-*
-* $Source: /dca/cvsroot/houdini/VRAY_clusterThis/VRAY_clusterThisRender.C,v $
-*
-* $Author: mstory $
-*
-* See Change History at the end of the file.
-*
-*    Digital Cinema Arts (C) 2008
-*
-* This work is licensed under the Creative Commons Attribution-ShareAlike 2.5 License.
-* To view a copy of this license, visit http://creativecommons.org/licenses/by-sa/2.5/ or send a letter to
-* Creative Commons, 543 Howard Street, 5th Floor, San Francisco, California, 94105, USA.
 *
 ***************************************************************************** */
 
@@ -334,7 +319,6 @@ void VRAY_clusterThis::render()
                if(myUseTempFile) {
                      cout << "VRAY_clusterThis::render() temp file: " << myTempFname << std::endl;
                      ofstream myGeoStream;
-                     // myGeoStream.open("/tmp/thisGeo.bgeo");
                      myGeoStream.open((const char *)myTempFname, ios_base::binary);
                      UT_Options myOptions;
                      inst_gdp->save(myGeoStream, 1, &myOptions);
